tuf_agent-backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Ahsan462aggk",email = "ahs462agk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "langchain (>=1.0.5,<2.0.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "tavily-python (>=0.7.12,<0.8.0)",
    "langchain-google-genai (>=3.0.3,<4.0.0)",
    "fastapi (>=0.121.1,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "piper-tts (>=1.3.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "websockets (>=15.0.1,<16.0.0)",
    "wsproto (>=1.3.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
